Revert `FormErrorMessage`
