--- packages/web-ui/src/FormErrorMessage/index.ts
@@ -1,2 +1,0 @@
-export { FormErrorMessage } from './FormErrorMessage';
-export type { FormErrorMessageProps } from './FormErrorMessage';
