+++ packages/web-ui/src/index.ts
@@ -12,7 +12,6 @@ export * from './Label';
 export * from './Fieldset';
 export * from './FieldsetLegend';
 export * from './FormHelperText';
-export * from './FormErrorMessage';
 export * from './Grid';
 export * from './Heading';
 export * from './Link';
